fix(router): roll back router
