+++ frontend/src/App.js
@@ -1,6 +1,14 @@
 import Component from "./core/Component.js";
-import router from "./core/router.js";
-import routes from "./core/routes.js";
+import Router from "./core/router.js";
+import LobbyPage from "./pages/LobbyPage.js";
+import LandingPage from "./pages/LandingPage.js";
+import MyPage from "./pages/MyPage.js";
+import ProfilePage from "./pages/ProfilePage.js";
+import PvPGamePage from "./pages/PvPGamePage.js";
+import SignUpPage from "./pages/SignUpPage.js";
+import TournamentGamePage from "./pages/TournamentGamePage.js";
+import JincparkTestPage from "./pages/JincparkTestPage.js";
+import HyeonjunTestPage from "./pages/HyeonjunTestPage.js";
 
 export default class App extends Component {
   template() {
@@ -10,6 +18,17 @@ export default class App extends Component {
   }
 
   mounted() {
-    router(routes);
+    const routes = {
+      "/": (element) => new LobbyPage(element),
+      "/landing": (element) => new LandingPage(element),
+      "/sign-up": (element) => new SignUpPage(element),
+      "/pvp": (element) => new PvPGamePage(element),
+      "/tournament": (element) => new TournamentGamePage(element),
+      "/my-page": (element) => new MyPage(element),
+      "/profile": (element) => new ProfilePage(element),
+      "/test-jincpark": (element) => new JincparkTestPage(element),
+      "/test-hyeonjun": (element) => new HyeonjunTestPage(element),
+    };
+    Router.init(routes);
   }
 }
